requirements.txt
Django
Pillow
django-tinymce
django-filebrowser
django-grappelli